fix: relationship context refactor
