+++ src/components/Relationship.tsx
@@ -1,17 +1,12 @@
 'use client';
 
 import styled from '@emotion/styled';
-import { useEffect, useRef, useState } from 'react';
 
 import { ConnectLine } from './ConnectLine';
 import { RelationshipMenu } from './RelationshipMenu';
 
 import { ERDRelation, ERDTable } from '@/features/erd-project';
-import {
-  Direction,
-  getStartEndDirection,
-  TableDirectionChild,
-} from '@/features/table-mapping';
+import { useContextMenu, useRelationshipData } from '@/features/mapping';
 
 interface RelationshipProps {
   tables: ERDTable[];
@@ -26,119 +21,15 @@ export function Relationship({
   deleteRelation,
   updateRelation,
 }: RelationshipProps) {
-  const relationDuplicate: ERDRelation[] = [];
-
-  const tableDirection: Map<ERDTable['id'], TableDirectionChild> = new Map();
-
-  const selfReferenceMapping: Map<ERDTable['id'], number> = new Map();
-
-  const [contextMenu, setContextMenu] = useState<{
-    x: number;
-    y: number;
-  } | null>(null);
-  const [lastRelation, setLastRelation] = useState<ERDRelation | null>(null);
-
-  const menuRef = useRef<HTMLDivElement | null>(null);
-
-  tables.forEach((table) => {
-    relations[table.id]?.forEach((relation) => {
-      if (!relationDuplicate.includes(relation))
-        relationDuplicate.push(relation);
-    });
-  });
-
-  tables.forEach((table) => {
-    tableDirection.set(
-      table.id,
-      new Map([
-        ['LEFT', []],
-        ['RIGHT', []],
-        ['TOP', []],
-        ['BOTTOM', []],
-      ]),
-    );
-
-    selfReferenceMapping.set(table.id, 0);
-  });
-
-  relationDuplicate.forEach((relation) => {
-    const fromTable = tables.find((t) => t.id === relation.from);
-    const toTable = tables.find((t) => t.id === relation.to);
-
-    if (fromTable && toTable && fromTable.id !== toTable.id) {
-      const { fromDirection, toDirection } = getStartEndDirection(
-        fromTable,
-        toTable,
-      );
-
-      tableDirection
-        .get(fromTable.id)
-        ?.get(fromDirection)
-        ?.push({
-          sortVal:
-            fromDirection === 'LEFT' || fromDirection === 'RIGHT'
-              ? toTable.top
-              : toTable.left,
-          tableID: toTable.id,
-          relationID: relation.id,
-        });
-
-      tableDirection
-        .get(toTable.id)
-        ?.get(toDirection)
-        ?.push({
-          sortVal:
-            toDirection === 'LEFT' || toDirection === 'RIGHT'
-              ? fromTable.top
-              : fromTable.left,
-          tableID: fromTable.id,
-          relationID: relation.id,
-        });
-    }
-  });
-
-  const directions: Direction[] = ['TOP', 'BOTTOM', 'LEFT', 'RIGHT'];
-
-  directions.forEach((direction) => {
-    tableDirection.keys().forEach((tableID) =>
-      tableDirection
-        .get(tableID)
-        ?.get(direction)
-        ?.sort((a, b) => a.sortVal - b.sortVal),
-    );
-  });
-
-  const handleContextMenu = (
-    event: React.MouseEvent,
-    relation: ERDRelation,
-  ) => {
-    event.preventDefault();
-    setContextMenu({ x: event.clientX, y: event.clientY });
-    setLastRelation(relation);
-  };
-
-  const closeContextMenu = () => {
-    setContextMenu(null);
-    setLastRelation(null);
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (e: MouseEvent) => {
-      if (!menuRef.current?.contains(e.target as Node)) {
-        closeContextMenu();
-      }
-    };
-
-    if (contextMenu) {
-      window.addEventListener('mousedown', handleClickOutside);
-    } else {
-      window.removeEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      window.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [contextMenu]);
+  const { relationDuplicate, tableDirection, selfReferenceMapping } =
+    useRelationshipData(tables, relations);
+  const {
+    contextMenu,
+    lastRelation,
+    menuRef,
+    openContextMenu,
+    closeContextMenu,
+  } = useContextMenu();
 
   return (
     <styles.display>
@@ -165,7 +56,7 @@ export function Relationship({
             relation={relation}
             tableDirection={tableDirection}
             selfReferenceMapping={selfReferenceMapping}
-            handleContextMenu={(e) => handleContextMenu(e, relation)}
+            handleContextMenu={(e) => openContextMenu(e, relation)}
           />
         ))}
       </styles.wrapper>
