+++ src/features/mapping/mapping.type.ts
