--- src/features/table-mapping/index.ts
@@ -1,2 +1,0 @@
-export * from './table-mapping.helper';
-export * from './table-mapping.type';
